Feat: User can delete from their outfits
